requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
pinecone-client==3.0.0
numpy==1.24.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6
langchain==0.0.340
langchain-openai==0.0.2
langchain-pinecone==0.0.0
tiktoken==0.5.2
